fix: update get price usd
